Patching up recently blocked

+++ lib/screens/settings_screen.dart
@@ -7,6 +7,7 @@ import 'package:flutter_hole/models/preferences/preference_is_dark.dart';
 import 'package:flutter_hole/models/preferences/preference_port.dart';
 import 'package:flutter_hole/models/preferences/preference_token.dart';
 import 'package:flutter_hole/models/preferences/setting_widget.dart';
+import 'package:fluttertoast/fluttertoast.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class SettingsScreen extends StatefulWidget {
@@ -15,6 +16,15 @@ class SettingsScreen extends StatefulWidget {
 }
 
 class _SettingsScreenState extends State<SettingsScreen> {
+  static int resetCount = 0;
+  static final int reset = 5;
+
+  @override
+  void dispose() {
+    resetCount = 0;
+    super.dispose();
+  }
+
   final List<Preference> prefList = [
     PreferenceIsDark(),
     PreferenceHostname(),
@@ -45,19 +55,28 @@ class _SettingsScreenState extends State<SettingsScreen> {
                   padding: const EdgeInsets.all(8.0),
                   child: RaisedButton(
                     onPressed: () {
-                      SharedPreferences.getInstance().then((preferences) {
-                        preferences.clear().then((didClear) {
-                          if (didClear) {
-                            for (Preference preference in prefList) {
-                              preference.set(context);
+                      if (++resetCount == reset) {
+                        resetCount = 0;
+                        SharedPreferences.getInstance().then((preferences) {
+                          preferences.clear().then((didClear) {
+                            if (didClear) {
+                              for (Preference preference in prefList) {
+                                preference.set(context);
+                              }
+                              Fluttertoast.showToast(msg: 'Factory reset');
+                              AppState.of(context).updateStatus();
+                              Navigator.pushReplacement(
+                                  context, MaterialPageRoute(
+                                  builder: (context) => SettingsScreen()));
+                            } else {
+                              print('no luck');
                             }
-                            print('yay cleared');
-                            AppState.of(context).updateStatus();
-                          } else {
-                            print('no luck');
-                          }
+                          });
                         });
-                      });
+                      } else {
+                        Fluttertoast.showToast(msg: 'Press ${reset -
+                            resetCount} more times to reset');
+                      }
                     },
                     child: Text('Reset to default settings'),
                     color: Colors.red,
